Rework TestWeb.Auth.authenticate
The old one was set up to be a function plug. Well, I wound up
dependency injecting TestWeb.Auth in all environments except test,
where I dependency inject a module that just lets the user
through. That system was incompatible with the way authenticate was
used, so I had to change it.